legados/sysava/requirements.txt
streamlit
supabase
bcrypt
python-dotenv
google-generativeai
google-auth-oauthlib
openai
pypdf
Pillow
requests
pyinstaller
streamlit-pdf-viewer
streamlit-javascript
pyinstaller
customtkinter
google-generativeai
google-auth-oauthlib
openai
pypdf
